requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
httpx                 # для async-запросов (например, в Aiogram)
aiogram
celery
python-dotenv
pyjwt
